set the filed of ObjectMap package private & formate code & rename parameter of constructor
